better interface for run_element
the `run_element` function now encapsulates the necessary information for analysis in its parameters. The `element` variable that it is given contains a "src" attribute on which there is a path that points to the element that should be analysed. It also contains a "dest" attribute on which there  is a path to the folder where the results of analysis should be saved.